:truck: move old selectors to deprecated state
we're still keeping them around for future reference and because we haven't achieved feature parity yet
